+++ snapatac2-core/src/preprocessing/mod.rs
@@ -6,6 +6,7 @@ mod qc;
 pub use bam::{make_fragment_file, BamQC, FlagStat};
 pub use import::{import_contacts, import_fragments, import_values, ChromValue};
 pub use qc::{
-    fraction_of_reads_in_region, fragment_size_distribution, get_barcode_count, make_promoter_map,
+    SummaryType,
+    get_barcode_count, make_promoter_map,
     read_tss, CellBarcode, Contact, Fragment, QualityControl, TSSe, TssRegions,
 };