+++ snapatac2-python/src/lib.rs
@@ -48,6 +48,7 @@ fn _snapatac2(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(preprocessing::tss_enrichment, m)?)?;
     m.add_function(wrap_pyfunction!(preprocessing::add_frip, m)?)?;
     m.add_function(wrap_pyfunction!(preprocessing::fragment_size_distribution, m)?)?;
+    m.add_function(wrap_pyfunction!(preprocessing::summary_by_chrom, m)?)?;
 
     m.add_function(wrap_pyfunction!(export::export_fragments, m)?)?;
     m.add_function(wrap_pyfunction!(export::export_coverage, m)?)?;